refactor(preview): extract preview layout helpers
